feat: 스터디룸 프리뷰 탭 추가 및 수업 상담하기 연결 - 프리뷰 콘텐츠 탭 구조 추가 (소개 요약 / 문의 내역) - 기존 소개 영역 intro-tab.tsx로 분리 - 문의 내역 탭: GET /public/study-rooms/{id}/inquiries 연동, 제목 입력 → 작성 페이지 이동 - 수업 상담하기 버튼 연결 (PUBLIC.CONSULTATION.CREATE) - consultation-write-area: studyRoomId select 초기값 세팅, sessionStorage 제목 pre-fill - StatusBadge: className prop 추가, active → primary 리네임
