+++ src/app/(public)/study-room-preview/[id]/[teacherId]/page.tsx
@@ -30,13 +30,11 @@ export default async function StudyRoomDetailPage({ params }: PageProps) {
               studyRoomId={studyRoomId}
             />
           </ColumnLayout.Left>
-          <ColumnLayout.Right className="desktop:max-w-[740px] desktop:px-8 flex flex-col rounded-[12px]">
-            <div className="border-line-line1 flex flex-col gap-9 rounded-[12px] border bg-white py-2">
-              <StudyroomPreviewContents
-                studyRoomId={studyRoomId}
-                teacherId={teacherIdNum}
-              />
-            </div>
+          <ColumnLayout.Right className="desktop:max-w-[740px]">
+            <StudyroomPreviewContents
+              studyRoomId={studyRoomId}
+              teacherId={teacherIdNum}
+            />
           </ColumnLayout.Right>
           <ScrollToTopButton />
         </div>
