+++ src/entities/study-room-preview/infrastructure/preview.repository.ts
@@ -2,7 +2,11 @@ import { api } from '@/shared/api';
 import { unwrapEnvelope } from '@/shared/lib/api-utils';
 
 import { domain } from '../core';
-import type { PreviewMain, PreviewSide } from '../types';
+import type {
+  PreviewConsultationList,
+  PreviewMain,
+  PreviewSide,
+} from '../types';
 
 const getPreviewSide = async (
   teacherId: number,
@@ -31,9 +35,21 @@ const getPreviewMain = async (studyRoomId: number): Promise<PreviewMain> => {
   return unwrapEnvelope(response, domain.main);
 };
 
+const getPreviewConsultations = async (
+  studyRoomId: number,
+  params: { page: number; size: number }
+): Promise<PreviewConsultationList> => {
+  const response = await api.public.get(
+    `/public/study-rooms/${studyRoomId}/inquiries`,
+    { params }
+  );
+  return unwrapEnvelope(response, domain.consultationList);
+};
+
 export const repository = {
   preview: {
     getPreviewSide,
     getPreviewMain,
+    getPreviewConsultations,
   },
 };
