+++ src/entities/study-room-preview/types/index.ts
@@ -12,6 +12,8 @@ export type PreviewCharacteristicResolvedContentDTO = z.infer<
   typeof dto.characteristicResolvedContent
 >;
 export type PreviewReviewItemDTO = z.infer<typeof dto.reviewItem>;
+export type PreviewConsultationItemDTO = z.infer<typeof dto.consultationItem>;
+export type PreviewConsultationListDTO = z.infer<typeof dto.consultationList>;
 
 // Main DTO types
 export type PreviewCoreDTO = z.infer<typeof dto.core>;
@@ -28,3 +30,5 @@ export type PreviewCore = z.infer<typeof domain.core>;
 export type PreviewMain = z.infer<typeof domain.main>;
 export type PreviewSideItem = z.infer<typeof domain.sideItem>;
 export type PreviewSide = z.infer<typeof domain.side>;
+export type PreviewConsultationItem = z.infer<typeof domain.consultationItem>;
+export type PreviewConsultationList = z.infer<typeof domain.consultationList>;
