+++ src/features/consultation/hooks/use-consultation-form.ts
@@ -1,18 +1,23 @@
 import { useRouter } from 'next/navigation';
 
 import { CreateConsultationPayload, repository } from '@/entities/consultation';
-import { useMutation } from '@tanstack/react-query';
+import { previewKeys } from '@/entities/study-room-preview';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 /**
  * [POST] 문의 작성
  */
 export function useCreateConsultation() {
   const router = useRouter();
+  const queryClient = useQueryClient();
 
   return useMutation({
     mutationFn: (params: CreateConsultationPayload) =>
       repository.createConsultation(params),
     onSuccess: (data) => {
+      queryClient.invalidateQueries({
+        queryKey: previewKeys.all,
+      });
       router.replace(`/consultation/${data.id}`);
     },
   });
